moved some logic into a new `getPopperOffsets` modifier
this allows autoPlacement to work properly
